requirements.txt
twilio>=9.8.5
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
